fix: SM-89: Python import loop is not the way
