--- src/models/base_model.py
@@ -1,4 +1,0 @@
-from sqlalchemy.orm import DeclarativeBase
-
-class Base(DeclarativeBase):
-    pass
